requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.23
Flask-Login==0.6.2
Flask-WTF==1.2.1
Flask-Migrate==4.0.5
Flask-Mail==0.9.1
Werkzeug==2.3.7
email-validator==2.1.0.post1
python-dotenv==1.0.0
matplotlib==3.7.2
openpyxl==3.1.2
